+++ robot/index.js
@@ -325,8 +325,8 @@ class Robot {
             await Apify.utils.puppeteer.blockRequests(page, this.options.trafficFilter);
 
         // const singleThread = setup.maxConcurrency === 1;
-        const shouldStartServer = !this.server && (prompt || (server && options.server.livecast.enable));
-        this.server = this.server || (shouldStartServer && startServer(page, setup, options.server.livecast));
+        const shouldStartServer = !this.server && (prompt || (server && options.server.interface.enable));
+        this.server = this.server || (shouldStartServer && startServer(page, setup, options.server.interface));
 
         decoratePage(this);
         initEventLogger(page, domain, input, options);
@@ -741,7 +741,7 @@ class Robot {
         }
 
         if (server) {
-            await sleep(options.server.livecast.snapshotTimeoutSecs);
+            await sleep(options.server.interface.snapshotTimeoutSecs);
             await server.serve(page);
             await sleep(5 * 1000);
         }