+++ robot/tools/hooks/index.js
@@ -74,7 +74,7 @@ const decoratePage = ({page, server}) => {
     LOGGER.triggerMethods.map(methodName => {
         const originalMethod = page[methodName];
 
-        if (server && SERVER.livecast.triggerMethods.some(livecastTrigger => methodName.includes(livecastTrigger))) {
+        if (server && SERVER.interface.triggerMethods.some(triggerMethod => methodName.includes(triggerMethod))) {
             page[methodName] = async (...args) => {
                 const argsForLog = args => args.map(arg => typeof arg === 'function' ? arg.toString().replace(/\s+/g, ' ') : arg);
                 console.log({[methodName]: argsForLog(args)});